# packages/server/server/common/api.yml
openapi: 3.0.1
info:
  title: server
  description: trecipe backend
  version: 1.0.0
servers:
- url: /api/v1
  description: Dev server
...
paths:
...
  /destinations:
    post:
      tags:
        - Destinations
      summary: Create new destination
      description: Create a new destination with given placeId
      requestBody:
        description: Google place id for this destination
        content:
          application/json:
            schema:
              type: object
              properties:
                placeId:
                  type: string
                  example: ChIJN1t_tDeuEmsRUsoyG83frY4
      responses:
        201:
          description: Returns the newly created destination
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Destination'
        500:
          $ref: '#/components/responses/InternalServerError'
    get:
      tags:
        - Destinations
      summary: Get destinations by trecipe uuid
      description: Get all destinations in a trecipe with specified uuid
      parameters:
        - $ref: '#/components/parameters/queryTrecipeId'
      responses:
        200:
          description: Returns an array of destinations in trecipe of given uuid
          content:
            application/json:
              schema:
                type: object
                properties:
                  destinations:
                    type: array
                    items:
                      $ref: '#/components/schemas/Destination'
...
components:
  parameters:
...
    queryTrecipeId:
      name: id
      in: query
      description: Trecipe uuid
      example: 123e4567-e89b-12d3-a456-426614174000
      required: true
      schema:
        type: string
        format: uuid
  schemas:
...
    Destination:
      title: Destination
      type: object
      properties:
        uuid:
          type: string
          format: uuid
          example: 2h357246-g725-4872-g84m-48692745878567
        name:
          type: string
          example: Sample destination
        category:
          type: array
          items:
            $ref: '#/components/schemas/DestinationCategory'
        geometry:
          $ref: '#/components/schemas/Geometry'
        formattedAddress:
          type: string
          example: 5, 48 Pirrama Rd, Pyrmont NSW 2009, Australia
        formattedPhoneNumber:
          type: string
          example: +61 2 9374 4000
        website:
          type: string
          example: https://www.google.com.au/about/careers/locations/sydney/
        rating:
          $ref: '#/components/schemas/Rating'
        userRatings:
          type: array
          items:
            $ref: '#/components/schemas/Rating'
          example: [4.2, 3.0, 4.5]
        placeId:
          type: string
          example: ChIJN1t_tDeuEmsRUsoyG83frY4
        photoRefs:
          type: array
          items:
            type: string
          example: [CnRtAAAATLZNl354RwP_9UKbQ_5Psy40texXePv4oAlgP4qNEkdIrkyse7rPXYGd9D_Uj1rVsQdWT4oRz4QrYAJNpFX7rzqqMlZw2h2E2y5IKMUZ7ouD_SlcHxYq1yL4KbKUv3qtWgTK0A6QbGh87GB3sscrHRIQiG2RrmU_jF4tENr9wGS_YxoUSSDrYjWmrNfeEHSGSc3FyhNLlBU]
    DestinationCategory:
      title: DestinationCategory
      type: string
      enum:
        - Food
        - Accommodation
        - Entertainment
        - Shopping
        - Sightseeing
    Geometry:
      title: Geometry
      type: object
      properties:
        lat:
          type: number
          format: double
          example: 35.929673
        lon:
          type: number
          format: double
          example: -78.948237
      required:
        - lat
        - lon
    Rating:
      title: Rating
      type: number
      format: float
      minimum: 0.0
      maximum: 5.0
      example: 4.3
...
    Error:
      title: Error
      type: object
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string
  responses:
...
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...